testing/web-platform/tests/css/css-color-adjust/meta/meta-color-scheme-remove.html

<!doctype html>
<title>CSS Color Adjust Test: remove meta tag</title>
<meta id="dark" name="color-scheme" content="dark">
<meta id="light" name="color-scheme" content="light">
<link rel="help" href="https://drafts.csswg.org/css-color-adjust-1/#color-scheme-meta">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<script>
  assert_root_color_scheme("dark");
  dark.remove();
  assert_root_color_scheme("light");
  light.remove();
  assert_root_color_scheme("normal");
</script>
